+++ src/Exception/InvalidJsonPointer.php
@@ -22,4 +22,12 @@ final class InvalidJsonPointer extends \InvalidArgumentException implements Exce
             $value,
         ));
     }
+
+    public static function fromUriFragmentIdentifierString(string $value): self
+    {
+        return new self(\sprintf(
+            'Value "%s" does not appear to be a valid URI fragment identifier representation of a JSON Pointer.',
+            $value,
+        ));
+    }
 }